+++ rust/guest_wrapper/artifacts/chain_guest/CHANGELOG.md
@@ -3,8 +3,10 @@
   * `e8175d54bac30389a87a8dd39ac43fe22f621066d4e657eee10a30328bdfb05f` – Downgrade revm from 19.2.0 to 19.0.0 [#1557](https://github.com/vlayer-xyz/vlayer/pull/1557)
   * `2dfd3adf61327dc1ad47a637f5771496458ad5dc9a10de41b05cb348fa5d2eee` – Use old guest IDs in chain guest [#1584](https://github.com/vlayer-xyz/vlayer/pull/1584)
   * `acbedbbeb664cdc1c58abd3d1c6c5a6e66d7cd206d752e3ddb7bf133f5f04ce7` – Use old guest IDs in chain host [#1603](https://github.com/vlayer-xyz/vlayer/pull/1603)
+<<<<<<< HEAD
   * `d35081d72b7d4875780f69da408127523669dfb2f17a05c23dc2506b74a92c83` – Revert Risc0 v1.2.1 to v1.2.0
   * `208d0580d2371c507fb6834b5d14c3b465b4049a0882834ae7e5a9767fd2ed3a` - Update local-testnet-op_spec anchor_state_registry [#1687](https://github.com/vlayer-xyz/vlayer/pull/1687)
+  * `cdfe3f19da5aed53ddde4d0c9f4297a1652d1712aee0e7990f80d89390ec9440` – Chain proof verifier refactor [#1701](https://github.com/vlayer-xyz/vlayer/pull/1701)
   * `9c027beee25f4f98a7cf995b5d752222673c6743e1cc6f38839b3bc6445f3441` – Initial teleport implementation [#1688](https://github.com/vlayer-xyz/vlayer/pull/1688)
   * `391fa348db0a76c8933fd780d5c4dcc72c2319986352ea8733e96eaa706f9d70` – Refactor optimism errors  * `2a556114dbe077cc7707573e55f2b646daa2bbd28e23bb3e7cc33876956e70f6` – Optimism teleport factories & clients
   * `a7f80dd271ebc7fb9e3f1c01640735dc4831a5de025532c35c1fd6405bd159fe` – Optimism teleport refactor